requirements.txt
PySide6>=6.7
SQLAlchemy>=2.0
pytest>=8.0
requests>=2.31
reportlab>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
psycopg[binary]>=3.1
